chore(analytics): update tsconfig.json
