+++ packages/analytics/tsconfig.json
@@ -20,5 +20,5 @@
     "allowUnreachableCode": false,
     "skipLibCheck": true
   },
-  "include": ["**/*.ts"],
+  "include": ["src/*"]
 }
